Release

+++ vent-soft/src/main/java/com/qualitysales/ventsoft/service/impl/ClientServiceImpl.java
@@ -1,6 +1,8 @@
 package com.qualitysales.ventsoft.service.impl;
 
 import com.qualitysales.ventsoft.Controllers.DTO.ClientDTO;
+import com.qualitysales.ventsoft.Controllers.DTO.ClientRequestDTO;
+import com.qualitysales.ventsoft.mapper.CityMapper;
 import com.qualitysales.ventsoft.mapper.ClientMapper;
 import com.qualitysales.ventsoft.model.City;
 import com.qualitysales.ventsoft.model.Client;
@@ -8,8 +10,13 @@ import com.qualitysales.ventsoft.repository.CityRepository;
 import com.qualitysales.ventsoft.repository.ClientRepository;
 import com.qualitysales.ventsoft.service.ClientService;
 import com.qualitysales.ventsoft.utils.HttpClientUtil;
+import com.qualitysales.ventsoft.utils.dto.GenericDTO;
+import com.qualitysales.ventsoft.utils.enums.MessagesEnum;
+import jakarta.persistence.EntityManager;
 import jakarta.transaction.Transactional;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.context.ApplicationContextException;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -20,6 +27,7 @@ import java.util.List;
 public class ClientServiceImpl implements ClientService {
 
     private final HttpClientUtil httpClientUtil;
+    private EntityManager entityManager;
 
     public ClientServiceImpl(ClientRepository clientRepository, HttpClientUtil httpClientUtil, CityRepository cityRepository) {
         this.clientRepository = clientRepository;
@@ -78,11 +86,18 @@ public class ClientServiceImpl implements ClientService {
 
     @Override
     public ClientDTO addClient(Client client) {
+        log.info("Client getDocTipe: {}",client);
+        City city = cityRepository.findById(client.getCity().getId()).orElseThrow(() -> new IllegalArgumentException("City not found"));
         try {
-            ClientDTO clientDTO = ClientMapper.MAPPER.toClient(client);
-            clientRepository.save(client);
-            log.info("addClient ok: {}", client);
-            return clientDTO;
+            if (city == null ) {
+                throw new IllegalArgumentException("City is null");
+            } else {
+                ClientDTO clientDTO = ClientMapper.MAPPER.toClient(client);
+                Client client1 = ClientMapper.MAPPER.toClientDTO(clientDTO);
+                clientRepository.save(client1);
+                log.info("addClient ok: {}", client1);
+                return clientDTO;
+            }
         } catch (Exception e) {
             log.error("addClient Error: {}", e.getMessage());
             throw new IllegalArgumentException(e);
@@ -90,24 +105,27 @@ public class ClientServiceImpl implements ClientService {
     }
 
     @Override
-    public Client updateClient(Integer id, ClientDTO clientDTO) {
+    public Client updateClient(Integer id, ClientRequestDTO clientrequestDTO) {
         Client idClient = clientRepository.findById(id).orElseThrow(() -> new IllegalArgumentException("Invalid client ID"));
         log.info("updateClient idClient ok: {}", idClient);
         City city = cityRepository.findById(idClient.getCity().getId()).orElseThrow(() -> new IllegalArgumentException("Invalid city ID"));
         log.info("updateClient city ok: {}", city);
+        City city1 = CityMapper.MAPPER.toCityDTO(clientrequestDTO.getCity());
+
         try {
-            idClient.setName(clientDTO.getName());
-            idClient.setLastName(clientDTO.getLastName());
-            idClient.setDocument(clientDTO.getDocument());
-            idClient.setCity(city);
-            idClient.setResidence(clientDTO.getResidence());
-            idClient.setCellPhone(clientDTO.getCellPhone());
-            idClient.setEmail(clientDTO.getEmail());
-            idClient.setEstate(clientDTO.getEstate());
+            idClient.setName(clientrequestDTO.getName());
+            idClient.setLastName(clientrequestDTO.getLastName());
+            idClient.setDocTipe(clientrequestDTO.getDocTipe());
+            idClient.setDocument(clientrequestDTO.getDocument());
+            idClient.setCity(city1);
+            idClient.setResidence(clientrequestDTO.getResidence());
+            idClient.setCellPhone(clientrequestDTO.getCellPhone());
+            idClient.setEmail(clientrequestDTO.getEmail());
+            idClient.setStatus(clientrequestDTO.getStatus());
 
             Client updatedClient = clientRepository.save(idClient);
 
-            log.info("updateClient ok: {}", clientDTO);
+            log.info("updateClient ok: {}", clientrequestDTO);
 
             return updatedClient;
 
@@ -118,14 +136,20 @@ public class ClientServiceImpl implements ClientService {
     }
 
     @Override
-    public void deleteClient(Integer id) {
+    public GenericDTO deleteClient(Integer id) {
         Client client = clientRepository.findById(id).orElseThrow(() -> new IllegalArgumentException("Invalid client ID"));
         try {
             log.info("deleteClient ok: {}", client);
-            clientRepository.deleteById(client.getId());
+            if (clientRepository.existsById(id)) {
+                clientRepository.deleteById(client.getId());
+                return GenericDTO.genericSuccess(MessagesEnum.REQUEST_SUCCESS, HttpStatus.OK.value());
+            }else {
+
+                throw new ApplicationContextException(MessagesEnum.REQUEST_FAILED.getMessage(), null);
+            }
         } catch (Exception e) {
             log.error("deleteClient Error: {}", e.getMessage());
-            throw new IllegalArgumentException(e);
+            throw new ApplicationContextException(e.getMessage());
         }
 
     }
